+++ d60.Cirqus/Views/ViewManagers/ViewInstanceUpdatedHandler.cs
@@ -1,0 +1,4 @@
+﻿namespace d60.Cirqus.Views.ViewManagers
+{
+    public delegate void ViewInstanceUpdatedHandler<TViewInstance>(TViewInstance instance) where TViewInstance : IViewInstance;
+}